+++ docs/sphinx/conf.py
@@ -24,7 +24,7 @@ extensions = [
    'sphinx.ext.doctest',
    'sphinx.ext.autodoc',
    'sphinx.ext.autosummary',
-    'sphinx.ext.napoleon',
+   'sphinx.ext.napoleon',
 ]
 
 autosummary_generate = True
