+++ tinytuya/BasicDevices/CoverDevice.py
@@ -5,69 +5,9 @@
 
  Author: Jason A. Cox
  For more information see https://github.com/jasonacox/tinytuya
-
- Local Control Classes
-    CoverDevice(...)
-        See OutletDevice() for constructor arguments
-
- Functions
-    CoverDevice:
-        open_cover(switch=None, nowait=False)       # Open the cover (switch defaults to DPS_INDEX_MOVE)
-        close_cover(switch=None, nowait=False)      # Close the cover (switch defaults to DPS_INDEX_MOVE)
-        stop_cover(switch=None, nowait=False)       # Stop the cover motion (switch defaults to DPS_INDEX_MOVE)
-        continue_cover(switch=None, nowait=False)   # Continue cover motion (if supported)
-        set_cover_type(cover_type)                  # Manually set cover type (1-8)
-
- Notes
-    CoverDevice automatically detects the device type (1-8) based on status response:
-    
-    Type 1: ["open", "close", "stop", "continue"] - Most curtains, blinds, roller shades (DEFAULT)
-    Type 2: [true, false]                         - Simple relays, garage doors, locks  
-    Type 3: ["0", "1", "2"]                       - String-numeric position/state
-    Type 4: ["00", "01", "02", "03"]              - Zero-prefixed numeric position/state
-    Type 5: ["fopen", "fclose"]                   - Directional binary (no stop)
-    Type 6: ["on", "off", "stop"]                 - Switch-lexicon open/close
-    Type 7: ["up", "down", "stop"]                - Vertical-motion (lifts, hoists)
-    Type 8: ["ZZ", "FZ", "STOP"]                  - Vendor-specific (Abalon-style, older standard)
-    
-    Credit for discovery: @make-all in https://github.com/jasonacox/tinytuya/issues/653
-    Detection occurs on first command by checking device status. Uses priority ordering
-    to handle overlapping values (Type 1 has highest priority). Defaults to Type 1 if
-    detection fails. You can manually override using set_cover_type(type_id) if needed.
-    
-    Common DPS IDs:
-    - DPS 1: Most common for cover control
-    - DPS 101: Second most common (often backlight or secondary function)
-    - DPS 4: Commonly used for second curtain in dual-curtain devices
-      (DPS 2 and 3 typically for position write/read, DPS 5 and 6 for second curtain,
-       with configuration and timers starting from DPS 7 onward)
-
-    Inherited
-        json = status()                    # returns json payload
-        set_version(version)               # 3.1 [default] or 3.3
-        set_socketPersistent(False/True)   # False [default] or True
-        set_socketNODELAY(False/True)      # False or True [default]
-        set_socketRetryLimit(integer)      # retry count limit [default 5]
-        set_socketTimeout(timeout)         # set connection timeout in seconds [default 5]
-        set_dpsUsed(dps_to_request)        # add data points (DPS) to request
-        add_dps_to_request(index)          # add data point (DPS) index set to None
-        set_retry(retry=True)              # retry if response payload is truncated
-        set_status(on, switch=1, nowait)   # Set status of switch to 'on' or 'off' (bool)
-        set_value(index, value, nowait)    # Set int value of any index.
-        heartbeat(nowait)                  # Send heartbeat to device
-        updatedps(index=[1], nowait)       # Send updatedps command to device
-        turn_on(switch=1, nowait)          # Turn on device / switch #
-        turn_off(switch=1, nowait)         # Turn off
-        set_timer(num_secs, nowait)        # Set timer for num_secs
-        set_debug(toggle, color)           # Activate verbose debugging output
-        set_sendWait(num_secs)             # Time to wait after sending commands before pulling response
-        detect_available_dps()             # Return list of DPS available from device
-        generate_payload(command, data)    # Generate TuyaMessage payload for command with data
-        send(payload)                      # Send payload to device (do not wait for response)
-        receive()
 """
 
-from .core import Device
+from ..core import Device
 
 class CoverDevice(Device):
     """
