+++ tinytuya/BasicDevices/__init__.py
@@ -1,0 +1,7 @@
+# TinyTuya Module
+# -*- coding: utf-8 -*-
+
+from .OutletDevice import OutletDevice
+from .CoverDevice import CoverDevice
+from .BulbDevice import BulbDevice
+
